requirement.txt
python-dotenv
langchain-google-genai
langchain-core
python-telegram-bot
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
